+++ modules/console_settings.py
@@ -1,0 +1,13 @@
+import sys
+import ctypes
+
+
+def set_window_title(detected_language, confidence):
+    title = f"Model: {model} - {detected_language} [{confidence:.2f}%]"
+
+    if sys.platform == "win32":
+        ctypes.windll.kernel32.SetConsoleTitleW(title)
+    else:
+        sys.stdout.write(f"\x1b]2;{title}\x1b\x5c")
+        sys.stdout.flush()
+
